API breaking changes for 4.0
CBL-1946 CBL-3813

Migrate configuration objects to be init only (much less checking required to ensure it is not changed while in use)
Fix a few potential null deferences based on static analysis
Undeprecate the BasicAuthenticator string constructor since there are no plans to ever remove it
Fix tests to use init only style
Fix an unused variable in C4ReplicatorWrapper

+++ src/Couchbase.Lite.Tests.Shared/LogTest.cs
@@ -116,7 +116,7 @@ namespace Test
                 #endif
 
                 Directory.EnumerateFiles(logDirectory)
-                    .Count().ShouldBe(totalCount, "because old log files should be getting pruned");
+                    .Count().ShouldBe((int)totalCount, "because old log files should be getting pruned");
             });
         }
